# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SGLCN-X20 Minting API",
    "description": "API for minting SeagullCoin-only NFTs, including buying and selling NFTs",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/mint": {
      "post": {
        "summary": "Mint an NFT with SeagullCoin",
        "operationId": "mintNFT",
        "parameters": [
          {
            "name": "nftMetadata",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NFTMetadata"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "NFT minted successfully"
          }
        }
      }
    },
    "/list-nft": {
      "post": {
        "summary": "List an NFT for sale",
        "operationId": "listNFT",
        "parameters": [
          {
            "name": "nftId",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NFTListing"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "NFT listed successfully"
          }
        }
      }
    },
    "/buy-nft": {
      "post": {
        "summary": "Buy an NFT with SeagullCoin",
        "operationId": "buyNFT",
        "parameters": [
          {
            "name": "nftId",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NFTPurchase"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "NFT bought successfully"
          }
        }
      }
    }
  },
  "definitions": {
    "NFTMetadata": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "file": {
          "type": "string"
        }
      }
    },
    "NFTListing": {
      "type": "object",
      "properties": {
        "nftId": {
          "type": "integer"
        },
        "price": {
          "type": "number"
        }
      }
    },
    "NFTPurchase": {
      "type": "object",
      "properties": {
        "nftId": {
          "type": "integer"
        },
        "price": {
          "type": "number"
        }
      }
    }
  }
}